Solve scaling issue, create gridRows for gridItems
To allow the grid to keep the same size container while being able to
take on any amount of grid items evenly inside, I needed to create this
new system:

gridContainer > gridRows > gridItems

The amount of rows is equal to the input of the user as well as is the
gridItems, so I first create a number of grid rows, then add the
same amount of gridItems, and it scales perfectly!

I also got rid of the grid item's borders for visual purposes.